Correct type of zoomSpeed
`zoomSpeed` is introduced for configurable manual zoom speed and the
type of value is `number` as described in [1].

However, it should be an array of `number` if it is used with
Configurator class, or an error is occured in _handleObject().
This update is to correct the type of zoomSpeed and fix #37.

[1] https://github.com/almende/vis/pull/3657