chore(examples): generate examples index
This generates single index file with (potentialy nested) groups and
generated screenshots of examples.

It's a bit rough around the edges but most of it should be easily solved
with a little CSS or similar tweaking. There are just 3 examples that
don't render well at the moment. Two don't render at all (everythings
white) and one doesn't render it's main feature (most likely because
it's not rendered on the canvas).

The whole thing is written to be reusable, therefore there should be no
problem porting this to Graph3D or Timeline. However it get's almost all
of it's information from the examples. All examples have to have the
canvas in a container with given id (configurable on CLI). Their titles
have to follow this structure:
`Group | [Another Group | [Another Group | …]] Example Title`.

There is also some primitive linting that requires the title to match
the heading in the body (not enforced in any way, just a warning).