chore(examples): keep stable layout on screenshots
This works by adding layout seed unless provided. It however fails when
the graph is randomly generated (we could eliminate this by ditching
Math.random() from exampleUtil in favor of deterministic generator) or
changes in time (the timing is not perfect but there aren't many of
these, maybe only one).