chore(examples): generate examples index (#140)
* chore(examples): generate examples index

This generates single index file with (potentialy nested) groups and
generated screenshots of examples.

It's a bit rough around the edges but most of it should be easily solved
with a little CSS or similar tweaking. There are just 3 examples that
don't render well at the moment. Two don't render at all (everythings
white) and one doesn't render it's main feature (most likely because
it's not rendered on the canvas).

The whole thing is written to be reusable, therefore there should be no
problem porting this to Graph3D or Timeline. However it get's almost all
of it's information from the examples. All examples have to have the
canvas in a container with given id (configurable on CLI). Their titles
have to follow this structure:
`Group | [Another Group | [Another Group | …]] Example Title`.

There is also some primitive linting that requires the title to match
the heading in the body (not enforced in any way, just a warning).

* chore(examples): add package script for examples index generation

* chore(examples): solve issues with examples

- All examples are in the index now.
- There are no empty screenshots.

PS: The Gephi example loads data from a server which is kinda impossible
without a server. It doesn't even work if it is opened over file://.

* chore(examples): update

* chore(examples): add JSFiddle

* chore(examples): install missing types

* chore(examples): update

* chore(examples): improve design

* chore(examples): add CodePen

Adding more code playgrounds will be really easy from now on.

* chore(examples): update

* chore(examples): move into scripts folder

* chore(examples): fix broken paths

* chore(examples): speed the animations up

* chore(examples): fix package script paths

* chore(examples): solve responsibility issues

* chore(examples): keep stable layout on screenshots

This works by adding layout seed unless provided. It however fails when
the graph is randomly generated (we could eliminate this by ditching
Math.random() from exampleUtil in favor of deterministic generator) or
changes in time (the timing is not perfect but there aren't many of
these, maybe only one).