fix(layout-engine): invalid hierarchical layout (#77)
* test(layout-engine): add failing test for this issue

* fix(layout-engine): generate valid hierarchical layouts

TODO: This is very slow at the moment, it needs to be optimized for
better performance in production.

* chore(dist): update

* fix(layout-engine): position directed hierarchy to the bottom

This mostly minimizes the amount of levels edges span between two nodes.

* chore(dist): update

* chore(layout-engine): remove unused code

* perf(layout-engine): improve level determination perf

In my tests up to 2500 % improvement in performance depending on the
size and structure of the graph.

* perf(layout-engine): use faster cycle detection

Cycles are detected as the levels are being filled. This get's rid of
the incredibly slow O(V²*E) cycle detection. Still uses the same
fallback logic for cyclic “hierarchies”.

* chore(dist): update

* test(layout-engine): add one more test case

* fix(layout-engine): ignore circular edges in hierarchical layout

* test(layout-engine): randomize edge order

* style(layout-engine): refomat new files

* chore(dist): update

* chore(dist): update