docs(i18n): add en/kr navigation
To the main index file and at the bottom of each docs file.
This also removes static file duplicities between docs.

PS: We should create a better system. This is very error prone with one
language, even more so with two and if we're eventually going to have
more it will soon become unbearable.