chore(examples): add clustering to hierarchy methods
The example is mostly rewritten. It has brand new design. A lot of weird
stuff was rewritten (like a comment talking about randomly generating
nodes when in fact it was just generating ordered numbers 0 to 18 in
simple for; edges being pushed one by one into an array; the String
constructor etc.)

The example now allows layouting method and shaking direction to be
chosen using radio buttons (ordinary buttons were used before) and
checkboxes to choose which nodes should be clustered.