fix(rng)!: use better seedable RNG (#352)
* fix(rng)!: use better seedable RNG

BREAKING CHANGE: The layout for the same graph and seed is different
when compared to the previous version.

This replaces the Math.sin() based ones and Math.random() with Alea
based RNG from Vis Util.

* test(physics): add simple pinning tests

This simply runs some basic physics configurations on a few sets of data
and tests that the positions of the nodes after stabilization didn't
change from the last time.

* test(snapshots): remove only

I had to change Math.pow to **. No idea why but without this change it
didn't want to output the same values. I just hope this is not going to
break in CI or other people's machines.