fix(configurator): fix bugs and document the format
This wasn't documented before and many people made various mistakes in
it. Some options were simply wrong so there are fixes for that. It also
includes a test to prevent some common mistakes, like setting the step
bigger than the range or initial value outside of the range, in the
future.