Don't override Node options with Group options
Node options are being overridden by the values defined by the Group,
which should be the case, since more specific options (node) should have
priority over more generic ones (group).

This was fixed for `color` [1], but there are still many useful options
such as borderWidth, image, and shape that are being overridden.

This patch fixes this by not merging from the group any new option that
has been set for the node.

Fix #408

[1]: https://github.com/almende/vis/issues/3713