fix: prevent reactitivy from messing with the state
This fixes issues with Vue's (and potentially other framework's)
reactivity cause weird and hard to debug errors in Vis Network.

There has already been multiple people confused by this.

One important thing though: What's the point of deleting all the modules
in destroy? I don't think it really does anything useful, it just breaks
the instance in pretty weird way. If I missed something though, this
really should be reconsidered as it's not exactly necessary to have
this, it's mostly just more convenient for the users.