fix(locale): use locale normalization (#737)
Prior to this it seems that some just added a bunch of possible values.
However that's not futureproof at all and become extremely tidious with
growing number of languages.

This PR adds a normalization function that is used to normalize the
input locale to something we can actually use. It takes multiple
possible formats into account (cs, cs-CZ, cs_CZ, CS_CZ, cs_cz, CS-cz…)
and converts them into standard format (cs-CZ or if we don't have that
variant then cs) that actually is in the list of locales we have or
fallbacks to English (en).