style: lint and fix changed files before commit
We already have this in some other repos. This will automatically fix
what can be fixed automatically (the comitter may not even notice) and
error out for changes that can't (these things would fail in CI anyway).

It can be skipped by `--no-verify` if necessary, though there should
never be the need to skip it.