fix(xss): don't use innerHTML to set text
We had a lot of instances of innerHTML that were simply setting text, in
quite a few cases parsing it as HTML was even counter productive as it
could theoreticly result in syntax errors and corrupted output (e.g. the
output of JSON.stringify definitely shouldn't be parsed as HTML).

There is only one occurence of innerHTML left and that is for the
node/edge title as the docs clearly state that any string passed will be
parsed as HTML.