refactor(activator): various changes (#1254)
Remove the dependency on Keycharm: All it really did was translate
`"esc"` to `event.keyCode === 27` (which is deprecated by the way). We
can just use `"key" in event ? event.key === "Escape" : event.keyCode
=== 27` instead of Keycharm and not relly solely on depracated APIs.

Remove the use of `addClassName` and `removeClassName`: These basically
polyfill IE9 and it's contemporaries (this is Windows Vista, iOS 4.3 and
Android Gingerbread era) which we don't otherwise polyfill so they don't
work one way or another.

Various other minor refactoring such as prefixing private properties
with underscore.